requirements.txt
django==1.11.20
psycopg2==2.7.3
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
django-filter==1.0.1
django-easy-pdf==0.1.0
xhtml2pdf==0.2b1
reportlab==3.3.0
django-logentry-admin==1.0.2
xlrd==1.1.0
